WIP: splitting away the device selector
